+++ Cargo.toml
@@ -36,12 +36,11 @@ quote = "1"
 proc-macro2 = "1"
 prettyplease = "0.1"
 toml = "0.5"
+tempdir = "0.3.7"
+once_cell = "1.7.2"
 
 [dev-dependencies]
 pgx-tests = "0.4.4"
-tempdir = "0.3.7"
-once_cell = "1.7.2"
-toml = "0.5.8"
 
 [profile.dev]
 panic = "unwind"
